Add and arrange icons
