+++ src/assets/icons.tsx
@@ -1,3 +1,5 @@
+import { MouseEventHandler } from "react";
+
 import { Icon } from "../styles";
 
 export const DeleteIcon = ({ onClick }: { onClick?: () => void }) => (
@@ -63,10 +65,10 @@ export const SortIcon = ({ onClick }: { onClick?: () => void }) => (
 	</svg>
 );
 
-export const More = ({ onClick }: { onClick: (id: string) => void }) => (
+export const Info = ({ onClick }: { onClick: MouseEventHandler<SVGSVGElement> }) => (
 	<Icon
 		xmlns="http://www.w3.org/2000/svg"
-		className="icon icon-tabler icon-tabler-dots-vertical"
+		className="icon icon-tabler icon-tabler-info-circle"
 		width="24"
 		height="24"
 		viewBox="0 0 24 24"
@@ -78,8 +80,8 @@ export const More = ({ onClick }: { onClick: (id: string) => void }) => (
 		onClick={onClick}
 	>
 		<path stroke="none" d="M0 0h24v24H0z" fill="none"></path>
-		<path d="M12 12m-1 0a1 1 0 1 0 2 0a1 1 0 1 0 -2 0"></path>
-		<path d="M12 19m-1 0a1 1 0 1 0 2 0a1 1 0 1 0 -2 0"></path>
-		<path d="M12 5m-1 0a1 1 0 1 0 2 0a1 1 0 1 0 -2 0"></path>
+		<path d="M3 12a9 9 0 1 0 18 0a9 9 0 0 0 -18 0"></path>
+		<path d="M12 9h.01"></path>
+		<path d="M11 12h1v4h1"></path>
 	</Icon>
 );
