+++ src/types/context.ts
@@ -1,18 +1,14 @@
-export interface IContact {
-	id: number;
-	avatar?: string;
-	email: string;
-	first_name: string;
-	phone: number;
-	[key: string]: string | undefined | React.ReactElement[] | number;
-}
+import { IContact } from "./contact";
 
 export type TSortOrder = "asc" | "desc";
 
 export type ContactContextType = {
 	contacts: IContact[];
 	getContacts: () => Promise<void>;
+	getContact: (id: number, contact: IContact[]) => Promise<void>;
+	selectedContact: IContact | null;
 	setContacts: (contacts: IContact[]) => void;
+	setSelectedContact?: (contact: IContact) => void;
 	setSortOrder: (order: TSortOrder) => void;
 	sortOrder: TSortOrder;
 	sortByName: () => void;
